bluetooth refactoring
