+++ bluedot/utils.py
@@ -5,6 +5,7 @@ import time
 SERVICE_NAME = "org.bluez"
 ADAPTER_INTERFACE = SERVICE_NAME + ".Adapter1"
 DEVICE_INTERFACE = SERVICE_NAME + ".Device1"
+PROFILE_MANAGER = SERVICE_NAME + ".ProfileManager1"
 
 def get_managed_objects():
     bus = dbus.SystemBus()
@@ -23,34 +24,27 @@ def find_adapter_in_objects(objects, pattern=None):
         if not pattern or pattern == adapter["Address"] or path.endswith(pattern):
             obj = bus.get_object(SERVICE_NAME, path)
             return dbus.Interface(obj, ADAPTER_INTERFACE)
-    raise Exception("Bluetooth adapter not found")
+    raise Exception("Bluetooth adapter {} not found".format(pattern))
+
+def get_adapter_property(device_name, property):
+    bus = dbus.SystemBus()
+    adapter_path = find_adapter(device_name).object_path
+    adapter = dbus.Interface(bus.get_object(SERVICE_NAME, adapter_path),"org.freedesktop.DBus.Properties")
+    return adapter.Get(ADAPTER_INTERFACE, property)    
 
 def get_mac(device_name):
-    bus = dbus.SystemBus()
-    adapter_path = find_adapter(device_name).object_path
-    adapter = dbus.Interface(bus.get_object(SERVICE_NAME, adapter_path),"org.freedesktop.DBus.Properties")
-    addr = adapter.Get(ADAPTER_INTERFACE, "Address")
-    return addr
+    return get_adapter_property(device_name, "Address")
 
 def get_adapter_powered_status(device_name):
-    bus = dbus.SystemBus()
-    adapter_path = find_adapter(device_name).object_path
-    adapter = dbus.Interface(bus.get_object(SERVICE_NAME, adapter_path),"org.freedesktop.DBus.Properties")
-    powered = adapter.Get(ADAPTER_INTERFACE, "Powered")
+    powered = get_adapter_property(device_name, "Powered")
     return True if powered else False
 
 def get_adapter_discoverable_status(device_name):
-    bus = dbus.SystemBus()
-    adapter_path = find_adapter(device_name).object_path
-    adapter = dbus.Interface(bus.get_object(SERVICE_NAME, adapter_path),"org.freedesktop.DBus.Properties")
-    discoverable = adapter.Get(ADAPTER_INTERFACE, "Discoverable")
+    discoverable = get_adapter_property(device_name, "Discoverable")
     return True if discoverable else False
 
 def get_adapter_pairable_status(device_name):
-    bus = dbus.SystemBus()
-    adapter_path = find_adapter(device_name).object_path
-    adapter = dbus.Interface(bus.get_object(SERVICE_NAME, adapter_path),"org.freedesktop.DBus.Properties")
-    pairable = adapter.Get(ADAPTER_INTERFACE, "Pairable")
+    pairable = get_adapter_property(device_name, "Pairable")
     return True if pairable else False
 
 def get_paired_devices(device_name):
@@ -92,6 +86,16 @@ def device_pairable(device_name, pairable):
         value = dbus.Boolean(0)
     adapter.Set(ADAPTER_INTERFACE, "Pairable", value)
 
+def device_powered(device_name, powered):
+    bus = dbus.SystemBus()
+    adapter_path = find_adapter(device_name).object_path
+    adapter = dbus.Interface(bus.get_object(SERVICE_NAME, adapter_path),"org.freedesktop.DBus.Properties")
+    if powered:
+        value = dbus.Boolean(1)
+    else:
+        value = dbus.Boolean(0)
+    adapter.Set(ADAPTER_INTERFACE, "Powered", value)
+
 def register_spp():
 
     service_record = """
@@ -122,7 +126,7 @@ def register_spp():
     """
 
     bus = dbus.SystemBus()
-    manager = dbus.Interface(bus.get_object(SERVICE_NAME, "/org/bluez"), "org.bluez.ProfileManager1")
+    manager = dbus.Interface(bus.get_object(SERVICE_NAME, "/org/bluez"), PROFILE_MANAGER)
 
     path = "/bluez"
     uuid = "00001101-0000-1000-8000-00805f9b34fb"
